mer rice, og starship prompt

+++ bin/configureStuff.sh
@@ -13,6 +13,6 @@ echo "Installing tmux plugin manager"
 [ ! -f "$HOME/.tmux/plugins/tpm/tpm" ] && git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm 
 
 echo "Symlinking stuff"
-sudo ln -s $BIN_DIR/addPac /usr/local/bin/addPac
-sudo ln -s $BIN_DIR/addAur /usr/local/bin/addAur
-sudo ln -s $BIN_DIR/../philemon /usr/local/bin/philemon
+[ ! -L /usr/local/bin/addPac ] && sudo ln -s $BIN_DIR/addPac /usr/local/bin/addPac
+[ ! -L /usr/local/bin/addAur ] && sudo ln -s $BIN_DIR/addAur /usr/local/bin/addAur
+[ ! -L /usr/local/bin/philemon ] && sudo ln -s $BIN_DIR/../philemon /usr/local/bin/philemon
